Rename stack-toast-controller.ts to toast-stack-controller.ts
